added convert tests
